工单07 颗粒度放宽 + 子目标折叠 + 浮点 round + 小面板 bug 修复
- 颗粒度：汇报/修正从 5 倍数放宽到任意正数 0.1%（issue 07 验收修订，庞大而简单的任务需要细粒度推进）；后端 percent f64 + valid_percent 1e-9 容差、前端 isValidPercentValue 同步容差（修 0.1/0.3/0.7 等一位小数严格等式误拒）
- 子目标列表折叠：更早已完成项折叠为「已完成 N 项」计数行 + 只展开最近一行（可撤销——撤销本就只能从末尾开始）；面板高度有界不出现滚动条
- 浮点四舍五入：后端 ProgressLog 写日志前 round_delta_minutes（1e-9）+ CurrentTaskView.percent 装配 round_to_one_decimal（一位小数）；前端 hoursFromMinutes / taskProgress 同步 round；解决「4.3100000000000005 / 10h」等 IEEE 754 浮点尾巴暴露到 UI；领域 TaskProgress::percent() 保持原始精度保护「分子不变分母变」语义
- 进度百分比显示口径：49.0 → "49"、49.3 → "49.3"（与 hoursFromMinutes 一致，去掉 Math.round 不再抹掉一位小数）；MiniBoardWindow 任务行 + PlanDetailPage 修正弹窗 + taskProgressLabel 三处统一
- 去 webview 滚动条：tokens.css body.transparent-root overflow:hidden 切断 html/body/#app 层滚动条；MiniBoardWindow .board-card 与 .picker 同步 overflow:hidden 兜底
- 修复有子目标任务误显百分比推进框：原 <ul v-if=has_subgoals> + <p v-if=remainingCount> + <div v-else> 三段切断了 v-else 链，v-else 实际绑到 remainingCount；用 <template v-if> 重组恢复完整「有子目标 / 无子目标」分支
- 文档同步：FUNC.md 前后端（isValidPercentStep → isValidPercentValue、hoursFromMinutes / taskProgress 描述、report_percent / correct_total 颗粒度、新增 round_delta_minutes / round_to_one_decimal 工具说明）、CONTEXT 进度百分比口径
- 测试：seam_progress 新增 progress_percent_rounds_to_one_decimal（边界 round 反复 0.1% 累计干净 43.1 + ProgressLog sum 接近 258.6 + 领域 percent() 原始精度保护）；全套 49 测通过

+++ src-tauri/src/commands.rs
@@ -161,12 +161,12 @@ pub fn undo_subgoal(
     ProgressService::undo_subgoal(&conn, state.clock.as_ref(), subgoal_id)
 }
 
-/// 无子目标任务增量汇报 +percent%（5–100 的 5 倍数，累计不超 100%）。
+/// 无子目标任务增量汇报 +percent%（任意正数，最小 0.1%、最多一位小数，累计不超 100%）。
 #[tauri::command]
 pub fn report_percent(
     state: State<'_, AppState>,
     task_id: i64,
-    percent: u32,
+    percent: f64,
 ) -> Result<(), PlanError> {
     let conn = state.db.lock().unwrap();
     ProgressService::report_percent(&conn, state.clock.as_ref(), task_id, percent).map(|_| ())
@@ -177,7 +177,7 @@ pub fn report_percent(
 pub fn correct_total_progress(
     state: State<'_, AppState>,
     task_id: i64,
-    percent: u32,
+    percent: f64,
 ) -> Result<(), PlanError> {
     let conn = state.db.lock().unwrap();
     ProgressService::correct_total(&conn, state.clock.as_ref(), task_id, percent).map(|_| ())
